more memory-efficient Batch()
The following steps were taken to reduce memory usage:

 * Compute stddev from moments in one step
 * Use autofunc.Sub and autofunc.Div
 * Remove mulAdd() call, which required a large intermediate

On large batches, we should observe that the consumed memory is cut in half. Whether or not we get this kind of gain has yet to be seen. I did run some preliminary memory benchmarks on a conv net, and there was some improvement, but only by 10% or so. However, a lot of things in a convnet use memory besides batchnorm layers.